# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "QUICKHIRE API",
    "description": "API Documentation for QUICKHIRE API. The world best hiring solution for developers!",
    "contact": {
      "name": "The Best Developer"
    }
  },
  "paths": {
...
    "/api/company/": {
      "get": {
        "tags": [
          "Company"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/job/": {
      "get": {
        "tags": [
          "Job"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Job"
        ],
        "description": "",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/job/company": {
      "get": {
        "tags": [
          "Job"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}